recheck_dns/requirements.txt
anyio==3.3.4
asgiref==3.4.1
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
dnspython==2.0.0
fastapi==0.70.0
psycopg2-binary==2.9.2
pydantic==1.8.2
pymongo==3.11.0
python-dotenv==0.19.2
PyYAML==6.0
requests==2.26.0
sniffio==1.2.0
starlette==0.16.0
typing_extensions==4.0.0
urllib3==1.26.7
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
Werkzeug==2.0.2
elasticsearch==7.9.1